requirements.txt
requests
beautifulsoup4
pdf2image
python-dotenv
